Move emscripten pthread into its own module
