+++ lib/emscripten/src/lib.rs
@@ -51,6 +51,7 @@ mod lock;
 mod math;
 mod memory;
 mod process;
+mod pthread;
 mod signal;
 mod storage;
 mod syscalls;
@@ -866,43 +867,43 @@ pub fn generate_emscripten_env(globals: &mut EmscriptenGlobals) -> ImportObject 
         "___cxa_free_exception" => func!(crate::emscripten_target::___cxa_free_exception),
         "___resumeException" => func!(crate::emscripten_target::___resumeException),
         "_dladdr" => func!(crate::emscripten_target::_dladdr),
-        "_pthread_attr_destroy" => func!(crate::emscripten_target::_pthread_attr_destroy),
-        "_pthread_attr_getstack" => func!(crate::emscripten_target::_pthread_attr_getstack),
-        "_pthread_attr_init" => func!(crate::emscripten_target::_pthread_attr_init),
-        "_pthread_attr_setstacksize" => func!(crate::emscripten_target::_pthread_attr_setstacksize),
-        "_pthread_cleanup_pop" => func!(crate::emscripten_target::_pthread_cleanup_pop),
-        "_pthread_cleanup_push" => func!(crate::emscripten_target::_pthread_cleanup_push),
-        "_pthread_cond_destroy" => func!(crate::emscripten_target::_pthread_cond_destroy),
-        "_pthread_cond_init" => func!(crate::emscripten_target::_pthread_cond_init),
-        "_pthread_cond_signal" => func!(crate::emscripten_target::_pthread_cond_signal),
-        "_pthread_cond_timedwait" => func!(crate::emscripten_target::_pthread_cond_timedwait),
-        "_pthread_cond_wait" => func!(crate::emscripten_target::_pthread_cond_wait),
-        "_pthread_condattr_destroy" => func!(crate::emscripten_target::_pthread_condattr_destroy),
-        "_pthread_condattr_init" => func!(crate::emscripten_target::_pthread_condattr_init),
-        "_pthread_condattr_setclock" => func!(crate::emscripten_target::_pthread_condattr_setclock),
-        "_pthread_create" => func!(crate::emscripten_target::_pthread_create),
-        "_pthread_detach" => func!(crate::emscripten_target::_pthread_detach),
-        "_pthread_equal" => func!(crate::emscripten_target::_pthread_equal),
-        "_pthread_exit" => func!(crate::emscripten_target::_pthread_exit),
-        "_pthread_self" => func!(crate::emscripten_target::_pthread_self),
-        "_pthread_getattr_np" => func!(crate::emscripten_target::_pthread_getattr_np),
-        "_pthread_getspecific" => func!(crate::emscripten_target::_pthread_getspecific),
-        "_pthread_join" => func!(crate::emscripten_target::_pthread_join),
-        "_pthread_key_create" => func!(crate::emscripten_target::_pthread_key_create),
-        "_pthread_mutex_destroy" => func!(crate::emscripten_target::_pthread_mutex_destroy),
-        "_pthread_mutex_init" => func!(crate::emscripten_target::_pthread_mutex_init),
-        "_pthread_mutexattr_destroy" => func!(crate::emscripten_target::_pthread_mutexattr_destroy),
-        "_pthread_mutexattr_init" => func!(crate::emscripten_target::_pthread_mutexattr_init),
-        "_pthread_mutexattr_settype" => func!(crate::emscripten_target::_pthread_mutexattr_settype),
-        "_pthread_once" => func!(crate::emscripten_target::_pthread_once),
-        "_pthread_rwlock_destroy" => func!(crate::emscripten_target::_pthread_rwlock_destroy),
-        "_pthread_rwlock_init" => func!(crate::emscripten_target::_pthread_rwlock_init),
-        "_pthread_rwlock_rdlock" => func!(crate::emscripten_target::_pthread_rwlock_rdlock),
-        "_pthread_rwlock_unlock" => func!(crate::emscripten_target::_pthread_rwlock_unlock),
-        "_pthread_rwlock_wrlock" => func!(crate::emscripten_target::_pthread_rwlock_wrlock),
-        "_pthread_setcancelstate" => func!(crate::emscripten_target::_pthread_setcancelstate),
-        "_pthread_setspecific" => func!(crate::emscripten_target::_pthread_setspecific),
-        "_pthread_sigmask" => func!(crate::emscripten_target::_pthread_sigmask),
+        "_pthread_attr_destroy" => func!(crate::pthread::_pthread_attr_destroy),
+        "_pthread_attr_getstack" => func!(crate::pthread::_pthread_attr_getstack),
+        "_pthread_attr_init" => func!(crate::pthread::_pthread_attr_init),
+        "_pthread_attr_setstacksize" => func!(crate::pthread::_pthread_attr_setstacksize),
+        "_pthread_cleanup_pop" => func!(crate::pthread::_pthread_cleanup_pop),
+        "_pthread_cleanup_push" => func!(crate::pthread::_pthread_cleanup_push),
+        "_pthread_cond_destroy" => func!(crate::pthread::_pthread_cond_destroy),
+        "_pthread_cond_init" => func!(crate::pthread::_pthread_cond_init),
+        "_pthread_cond_signal" => func!(crate::pthread::_pthread_cond_signal),
+        "_pthread_cond_timedwait" => func!(crate::pthread::_pthread_cond_timedwait),
+        "_pthread_cond_wait" => func!(crate::pthread::_pthread_cond_wait),
+        "_pthread_condattr_destroy" => func!(crate::pthread::_pthread_condattr_destroy),
+        "_pthread_condattr_init" => func!(crate::pthread::_pthread_condattr_init),
+        "_pthread_condattr_setclock" => func!(crate::pthread::_pthread_condattr_setclock),
+        "_pthread_create" => func!(crate::pthread::_pthread_create),
+        "_pthread_detach" => func!(crate::pthread::_pthread_detach),
+        "_pthread_equal" => func!(crate::pthread::_pthread_equal),
+        "_pthread_exit" => func!(crate::pthread::_pthread_exit),
+        "_pthread_self" => func!(crate::pthread::_pthread_self),
+        "_pthread_getattr_np" => func!(crate::pthread::_pthread_getattr_np),
+        "_pthread_getspecific" => func!(crate::pthread::_pthread_getspecific),
+        "_pthread_join" => func!(crate::pthread::_pthread_join),
+        "_pthread_key_create" => func!(crate::pthread::_pthread_key_create),
+        "_pthread_mutex_destroy" => func!(crate::pthread::_pthread_mutex_destroy),
+        "_pthread_mutex_init" => func!(crate::pthread::_pthread_mutex_init),
+        "_pthread_mutexattr_destroy" => func!(crate::pthread::_pthread_mutexattr_destroy),
+        "_pthread_mutexattr_init" => func!(crate::pthread::_pthread_mutexattr_init),
+        "_pthread_mutexattr_settype" => func!(crate::pthread::_pthread_mutexattr_settype),
+        "_pthread_once" => func!(crate::pthread::_pthread_once),
+        "_pthread_rwlock_destroy" => func!(crate::pthread::_pthread_rwlock_destroy),
+        "_pthread_rwlock_init" => func!(crate::pthread::_pthread_rwlock_init),
+        "_pthread_rwlock_rdlock" => func!(crate::pthread::_pthread_rwlock_rdlock),
+        "_pthread_rwlock_unlock" => func!(crate::pthread::_pthread_rwlock_unlock),
+        "_pthread_rwlock_wrlock" => func!(crate::pthread::_pthread_rwlock_wrlock),
+        "_pthread_setcancelstate" => func!(crate::pthread::_pthread_setcancelstate),
+        "_pthread_setspecific" => func!(crate::pthread::_pthread_setspecific),
+        "_pthread_sigmask" => func!(crate::pthread::_pthread_sigmask),
         "___gxx_personality_v0" => func!(crate::emscripten_target::___gxx_personality_v0),
         "_gai_strerror" => func!(crate::env::_gai_strerror),
         "_getdtablesize" => func!(crate::emscripten_target::_getdtablesize),
